Fix combat precision loss and improve enemy health UI
- Refactored Enemy, Stall, and Projectile models to use Float for health and damage.
- Updated MainViewModel and StallUpgradeManager to maintain float precision in combat calculations.
- Improved rendering z-sorting in GameBoard.kt to keep overhead effects above ground entities.
- Added percentage health text display to enemy health bars.
- Updated unit tests and documentation (AGENTS.md, fixes.md).

+++ app/src/main/java/com/messark/hawker/MainViewModel.kt
@@ -266,7 +266,7 @@ class MainViewModel @JvmOverloads constructor(
         return enemyList.shuffled(random)
     }
 
-    private fun getEnemyHP(type: EnemyType, wave: Int): Int {
+    private fun getEnemyHP(type: EnemyType, wave: Int): Float {
         return EnemyRegistry.get(type).getHp(wave)
     }
 
@@ -708,7 +708,7 @@ class MainViewModel @JvmOverloads constructor(
                 // Temporarily boost stall for fire calculation
                 val boostedStall = if (boost > 1.0f) {
                     stall.copy(
-                        damage = (stall.damage * boost).toInt(),
+                        damage = stall.damage * boost,
                         effectDurationMs = (stall.effectDurationMs * boost).toLong(),
                         freezeDurationMs = (stall.freezeDurationMs * boost).toLong()
                     )
@@ -860,7 +860,7 @@ class MainViewModel @JvmOverloads constructor(
                 null // Enemy is dead
             } else {
                 enemy.copy(
-                    health = currentHealth.toInt(),
+                    health = currentHealth,
                     freezeDurationMs = maxFreezeDuration,
                     speedBoostDurationMs = speedBoostDuration
                 )
@@ -1359,7 +1359,7 @@ class MainViewModel @JvmOverloads constructor(
 
     private fun calculateStatBoost(coord: AxialCoordinate, hexes: Map<AxialCoordinate, HexTile>): BoostResult {
         val adjacentCoords = GridUtils.getNeighbors(coord)
-        var totalBoostPercent = 0
+        var totalBoostPercent = 0f
         val providers = mutableListOf<AxialCoordinate>()
         adjacentCoords.forEach { adj ->
             val stall = hexes[adj]?.stall
